fix(docs): Kbar change theme
