+++ apps/docs/src/components/index.ts
@@ -28,6 +28,7 @@ export { default as ExampleBlock } from './templates/example-block';
 export { default as ThemeToggle } from './theme-toggle';
 export { default as ThemeSwitch } from './theme-switch';
 export { default as KBar } from './kbar';
+export { default as KBarWrapper } from './kbar-wrapper';
 export { default as Blockholder } from './blockholder';
 export { default as LooperBG } from './looper-bg';
 export { default as FeaturesGrid } from './features-grid';
