+++ apps/docs/src/pages/_app.tsx
@@ -1,21 +1,19 @@
 import React from 'react';
 import { NextPage } from 'next';
-import dynamic from 'next/dynamic';
 import { debounce } from 'lodash';
-import { NextRouter, Router, useRouter } from 'next/router';
+import { NextRouter, Router } from 'next/router';
 import { NextUIProvider } from '@nextui-org/react';
 import { ThemeProvider as NextThemesProvider } from 'next-themes';
 import NProgress from 'nprogress';
 import PlausibleProvider from 'next-plausible';
 import { AppInitialProps } from 'next/app';
 import { NextComponent } from '@lib/types';
-import generateKbarActions from '@lib/kbar-actions';
-import { KBarProvider } from 'kbar';
 import { lightTheme, darkTheme } from '../theme/shared';
 import { isProd } from '@utils/index';
 import RouterEvents from '@lib/router-events';
 import globalStyles from '../styles/globalStyles';
 import '../styles/sandpack.css';
+import { KBarWrapper as KBarProvider } from '@components';
 
 type AppPropsType<
   R extends NextRouter = NextRouter,
@@ -28,10 +26,6 @@ type AppPropsType<
 };
 
 type AppProps<P = {}> = AppPropsType<Router, P>;
-
-const KbarComponent = dynamic(() => import('../components/kbar'), {
-  ssr: false
-});
 
 NProgress.configure({ parent: '#app-container' });
 
@@ -48,8 +42,6 @@ RouterEvents.on('routeChangeError', () => {
 });
 
 const Application: NextPage<AppProps<{}>> = ({ Component, pageProps }) => {
-  const router = useRouter();
-  const kbarActions = generateKbarActions(router);
   globalStyles();
   return (
     <NextThemesProvider
@@ -62,16 +54,7 @@ const Application: NextPage<AppProps<{}>> = ({ Component, pageProps }) => {
     >
       <NextUIProvider>
         <PlausibleProvider domain="nextui.org" enabled={isProd}>
-          <KBarProvider
-            actions={kbarActions}
-            options={{
-              animations: {
-                enterMs: 250,
-                exitMs: 100
-              }
-            }}
-          >
-            <KbarComponent />
+          <KBarProvider>
             <Component {...pageProps} />
           </KBarProvider>
         </PlausibleProvider>
